Issue: WSW-23577 Display throughput in iperf report as a float.
[iperf report]
    UDP_DONE_SERVER (RX)
    Local address : 192.168.0.120  Port 5001
    Remote address : 192.168.0.40  Port 45186
    Bytes Transferred 264600
    Duration (ms) 2029
    Bandwidth (Mbitpsec) 1.04

+++ nw_utils/iperf.c
@@ -121,7 +121,7 @@ static void lwiperf_report(void *arg,
             (void)PRINTF(" Port %d \r\n", remote_port);
             (void)PRINTF(" Bytes Transferred %llu \r\n", bytes_transferred);
             (void)PRINTF(" Duration (ms) %d \r\n", ms_duration);
-            (void)PRINTF(" Bandwidth (Mbitpsec) %d \r\n", bandwidth_kbitpsec / 1000U);
+            (void)PRINTF(" Bandwidth (Mbitpsec) %.2f \r\n", (double)bandwidth_kbitpsec / 1000U);
         }
     }
     else
